fix: silence deprecation warnings on Ubuntu (wx enum-OR, ayatana pragma, IPV4 endpoint copy-assign)
Clears the 12 deprecation warnings a -Wdeprecated-declarations
-Wdeprecated-copy -Wdeprecated audit build emits on Ubuntu 26.04
(GCC). All 12 hits are in our own code, no wx / cryptopp / glib
noise. After this the same audit reports 0 warnings on Ubuntu, so
promoting -Werror=deprecated-* in CI becomes an option (the macOS
side still needs a cryptopp include-path cleanup before that flag
can go on globally).

8 x -Wdeprecated-enum-enum-conversion (C++20). Canonical wx sizer
idioms like wxEXPAND | wxALL OR two enums of different types
(wxStretch, wxDirection, wxAlignment). C++20 deprecated arithmetic
between distinct enum types. Fixed by casting the first operand to
int. Sites: src/amule.cpp, src/FirstRunWizard.cpp.

1 x -Wdeprecated-declarations in MuleTrayIcon.cpp. app_indicator_new
is marked deprecated by newer libayatana-appindicator headers because
upstream AyatanaIndicators split the library into libayatana-
appindicator-glib (GLib-only, GMenu-based). Migrating today would
lock out every mainstream distro since the -glib package isn't in
Ubuntu / Fedora / openSUSE / Debian repos yet. Wrapped the single
call site in a #pragma GCC diagnostic ignored push/pop pair with an
explanatory comment.

2 x -Wdeprecated-copy in LibSocketAsio.cpp. CamuleIPV4Endpoint has a
user-provided copy ctor, which under C++11+ makes the compiler-
generated copy-assignment operator deprecated. Fixed by adding an
explicit = default copy-assign right after the copy ctor.
Functionally identical to what the compiler was already generating.

+++ src/FirstRunWizard.cpp
@@ -288,7 +288,7 @@ wxWizardPageSimple *CFirstRunWizard::BuildConnectionPage()
 	}
 	m_speedCtrl->Append(_("Other / set manually"));
 	m_speedCtrl->SetSelection((int)s_profileCount); // "Other" until a profile is chosen
-	sizer->Add(m_speedCtrl, 0, wxEXPAND | wxBOTTOM, 12);
+	sizer->Add(m_speedCtrl, 0, static_cast<int>(wxEXPAND) | wxBOTTOM, 12);
 
 	wxFlexGridSizer *grid = new wxFlexGridSizer(2, 2, 6, 8);
 	grid->AddGrowableCol(1);
@@ -309,7 +309,7 @@ wxWizardPageSimple *CFirstRunWizard::BuildConnectionPage()
 	m_downloadCtrl->SetValue(thePrefs::GetMaxDownload());
 	grid->Add(m_downloadCtrl, 0, wxEXPAND);
 
-	sizer->Add(grid, 0, wxEXPAND | wxBOTTOM, 12);
+	sizer->Add(grid, 0, static_cast<int>(wxEXPAND) | wxBOTTOM, 12);
 
 	m_derivedLabel = new wxStaticText(page, wxID_ANY, wxEmptyString);
 	sizer->Add(m_derivedLabel, 0, wxEXPAND);
@@ -368,10 +368,10 @@ wxWizardPageSimple *CFirstRunWizard::BuildNetworkPage()
 	m_udpPortCtrl->SetValue(thePrefs::GetUDPPort());
 	grid->Add(m_udpPortCtrl, 0, wxEXPAND);
 
-	sizer->Add(grid, 0, wxEXPAND | wxBOTTOM, 12);
+	sizer->Add(grid, 0, static_cast<int>(wxEXPAND) | wxBOTTOM, 12);
 
 #ifdef ENABLE_UPNP
-	sizer->Add(new wxStaticLine(page), 0, wxEXPAND | wxBOTTOM, 8);
+	sizer->Add(new wxStaticLine(page), 0, static_cast<int>(wxEXPAND) | wxBOTTOM, 8);
 	m_upnpCtrl = new wxCheckBox(page, wxID_ANY, _("Enable UPnP for router port forwarding"));
 	m_upnpCtrl->SetValue(true);
 	sizer->Add(m_upnpCtrl, 0, wxBOTTOM, 4);
@@ -511,9 +511,9 @@ wxWizardPageSimple *CFirstRunWizard::BuildFoldersPage()
 	sizer->Add(new wxStaticText(page, wxID_ANY, _("Destination folder for downloads")), 0, wxBOTTOM, 4);
 	wxBoxSizer *incRow = new wxBoxSizer(wxHORIZONTAL);
 	m_incomingCtrl = new wxTextCtrl(page, wxID_ANY, thePrefs::GetIncomingDir().GetRaw());
-	incRow->Add(m_incomingCtrl, 1, wxALIGN_CENTER_VERTICAL | wxRIGHT, 6);
+	incRow->Add(m_incomingCtrl, 1, static_cast<int>(wxALIGN_CENTER_VERTICAL) | wxRIGHT, 6);
 	incRow->Add(new wxButton(page, ID_FRW_BROWSE_INCOMING, _("Browse")), 0);
-	sizer->Add(incRow, 0, wxEXPAND | wxBOTTOM, 4);
+	sizer->Add(incRow, 0, static_cast<int>(wxEXPAND) | wxBOTTOM, 4);
 
 	// Warn that everything dropped in the Incoming folder is shared, reusing
 	// the same hint shown on the Directories preferences panel.
@@ -527,7 +527,7 @@ wxWizardPageSimple *CFirstRunWizard::BuildFoldersPage()
 		new wxStaticText(page, wxID_ANY, _("Folder for temporary download files")), 0, wxBOTTOM, 4);
 	wxBoxSizer *tmpRow = new wxBoxSizer(wxHORIZONTAL);
 	m_tempCtrl = new wxTextCtrl(page, wxID_ANY, thePrefs::GetTempDir().GetRaw());
-	tmpRow->Add(m_tempCtrl, 1, wxALIGN_CENTER_VERTICAL | wxRIGHT, 6);
+	tmpRow->Add(m_tempCtrl, 1, static_cast<int>(wxALIGN_CENTER_VERTICAL) | wxRIGHT, 6);
 	tmpRow->Add(new wxButton(page, ID_FRW_BROWSE_TEMP, _("Browse")), 0);
 	sizer->Add(tmpRow, 0, wxEXPAND);
